requirements.txt
django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
ipython==8.30.0
python-dotenv==1.0.1